rabu1

+++ menu/catatan_perjalanan.php
@@ -36,7 +36,7 @@
                         <td><?= $value['lokasi'] ?></td>
                         <td><?= $value['suhu'] ?></td> 
                         <td class="text-center"><a href="update.php?id_catatan=<?= $value['id_catatan']; ?>"><button type="submit" class="btn btn-warning"><i class="fa fa-pen"></i> </button></a>
-                        <a href="hapus.php?id_catatan=<?= $value['id_catatan']; ?>"><button type="submit" class="btn btn-danger"><i class="fa fa-trash"></i></button></a></td>
+                        <a onclick="return confirm('Yakin ingin menghapus data?')" href="hapus.php?id_catatan=<?= $value['id_catatan']; ?> "><button type="submit" class="btn btn-danger"><i class="fa fa-trash"></i></button></a></td>
                     </tr>
                     <?php } ?>
                 </tbody>
@@ -44,4 +44,3 @@
         </div>
     </div>
     </div>
-</div>
